test: separate reviewer test
